requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
httpx
loguru
beautifulsoup4
stripe
psycopg2-binary
python-dotenv
email-validator
argon2-cffi
